refactor the search algorithm to make it more readable
